Move StopArea form files out of map into form "package"
Move files related to Stop Area creation/editing modals and forms into
the "components/forms/stop-area" directory from previous
"components/map/stop-areas" directory.
