+++ ui/src/components/forms/stop-area/ConfirmStopAreaDeletionDialog.tsx
