+++ ui/src/components/map/stop-areas/EditStopAreaModal.tsx
@@ -1,10 +1,10 @@
 import { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { submitFormByRef } from '../../../utils';
+import { StopAreaFormState } from '../../forms/stop-area';
 import { CustomOverlay } from '../CustomOverlay';
 import { Modal } from '../modal/Modal';
 import { StopAreaForm } from './StopAreaForm';
-import { StopAreaFormState } from './stopAreaFormSchema';
 
 const testIds = {
   modal: 'EditStopAreaModal',
